+++ app/src/main/java/com/example/walklist/utils/Models.kt
@@ -7,7 +7,7 @@ import java.util.*
 /**
  * A walk object representing a walk created by users.
  */
-data class Walk(val id: Int?, val title: String, var distance: Double, var duration: Double,
+data class Walk(val id: Int?, val title: String, var distance: Int, var duration: Double,
                 @SerializedName("started_at") var startedAt: Date,
                 @SerializedName("start_point_lat") val startPointLat: Double,
                 @SerializedName("start_point_long") val startPointLong: Double,
@@ -20,7 +20,7 @@ data class Walk(val id: Int?, val title: String, var distance: Double, var durat
                 @SerializedName("encoded_route") var encodedRoute: String) {
 
     constructor(title: String, startPointLat: Double, startPointLong: Double) :
-            this(null, title, 0.0, 0.0, Date(), startPointLat, startPointLong, null, null, null,
+            this(null, title, 0, 0.0, Date(), startPointLat, startPointLong, null, null, null,
                 Date(), startPointLat, startPointLong, PolyUtils.encode(listOf(LatLng(startPointLat, startPointLong))))
 
     override fun toString(): String = title
